Added the GA5 toolchain
This toolchain isn't fetched by buildroot, but it is necessary in order to
build GA5 and more recent images. As it may not always be available on the
Laird site, adding it here.

You can use it by placing it in /usr/local and adding it's bin directory to
your path. A "local" install can also happen if you place it somewhere in
your home-directory tree.